// src/app/enums/currencty-converter/currency-code-enum.ts
export enum CurrencyCode {

    MDL = "MDL",
    EUR = "EUR",
    USD = "USD",
    RUB = "RUB",
    RON = "RON",
    UAH = "UAH",
    AED = "AED",
    ALL = "ALL",
    AMD = "AMD",
    AUD = "AUD",
    AZN = "AZN",
    BGN = "BGN",
    BYN = "BYN",
    CAD = "CAD",
    CHF = "CHF",
    CNY = "CNY",
    CZK = "CZK",
    DKK = "DKK",
    GBP = "GBP",
    GEL = "GEL",
    HKD = "HKD",
    HUF = "HUF",
    ILS = "ILS",
    INR = "INR",
    ISK = "ISK",
    JPY = "JPY",
    KGS = "KGS",
    KRW = "KRW",
    KWD = "KWD",
    KZT = "KZT",
    MKD = "MKD",
    MYR = "MYR",
    NOK = "NOK",
    NZD = "NZD",
    PLN = "PLN",
    RSD = "RSD",
    SEK = "SEK",
    TJS = "TJS",
    TMT = "TMT",
    TRY = "TRY",
    UZS = "UZS",
    XDR = "XDR",
    BYR = "BYR",
    EEK = "EEK",
    HRK = "HRK",
    LTL = "LTL",
    SKK = "SKK",
    LVL = "LVL"
}